third_party/WebKit/LayoutTests/inspector/tracing/timeline-animation-frame-expected.txt

Tests the Timeline events for Animation Frame feature

RequestAnimationFrame Properties:
{
    data : {
        frame : <string>
        id : <number>
        stackTrace : <object>
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    thread : <string>
    type : "RequestAnimationFrame"
}
Text details for RequestAnimationFrame: timeline-animation-frame.html:8
FireAnimationFrame Properties:
{
    children : <object>
    data : {
        frame : <string>
        id : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "FireAnimationFrame"
}
Text details for FireAnimationFrame: timeline-animation-frame.html:8
CancelAnimationFrame Properties:
{
    data : {
        frame : <string>
        id : <number>
        stackTrace : <object>
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    thread : <string>
    type : "CancelAnimationFrame"
}
Text details for CancelAnimationFrame: timeline-animation-frame.html:11
